+++ src/time_evolution/ssesolve.jl
@@ -262,7 +262,7 @@ function ssesolveEnsembleProblem(
 
     ensemble_prob = TimeEvolutionProblem(
         EnsembleProblem(prob_sme, prob_func = _prob_func, output_func = _output_func[1], safetycopy = true),
-        prob_sme.tlist,
+        prob_sme.times,
         prob_sme.dimensions,
         (progr = _output_func[2], channel = _output_func[3]),
     )
@@ -418,8 +418,8 @@ function ssesolve(
 
     return TimeEvolutionStochasticSol(
         ntraj,
+        ens_prob.times,
         _sol_1.t,
-        ens_prob.tlist,
         states,
         expvals,
         expvals, # This is average_expect